Moved rcVector unit tests to their own file
